+++ src/pages/FavoriteRecipes/index.jsx
@@ -1,7 +1,4 @@
 import Typography from '@mui/material/Typography'
-import { orange, red } from '@mui/material/colors'
-import FavoriteIcon from '@mui/icons-material/Favorite'
-import ShareIcon from '@mui/icons-material/Share'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import {
   MoreVert as MoreVertIcon,
@@ -20,28 +17,27 @@ import {
   Container,
   Grid,
   IconButton,
+  Link,
   Tooltip,
   styled,
   useMediaQuery,
 } from '@mui/material'
 import { useContext, useEffect, useState } from 'react'
-const ExpandMore = styled((props) => {
-  const { expand, ...other } = props
-  return <IconButton {...other} />
+export const ExpandMore = styled((props) => {
+  const { expand, ...other } = props;
+  return <IconButton {...other} />;
 })(({ theme, expand }) => ({
   transform: !expand ? 'rotate(0deg)' : 'rotate(180deg)',
   marginLeft: 'auto',
   transition: theme.transitions.create('transform', {
     duration: theme.transitions.duration.shortest,
   }),
-}))
+}));
 import { useTheme } from '@mui/material/styles';
-
 import * as Tag from './index.js'
 import { api } from '../../api/index.js'
 import { AuthContext } from '../../contexts/AuthContext.jsx'
 import { INTERFACE } from '../INTERFACE/index.jsx'
-
 export const FavoriteRecipes = () => {
   const [favoriteRecipes, setFavoriteRecipes] = useState([])
   const { user } = useContext(AuthContext)
@@ -63,7 +59,6 @@ export const FavoriteRecipes = () => {
       console.error(error)
     }
   }
-
   useEffect(() => {
     const fetchFavoriteRecipes = async () => {
       const userId = user.uid
@@ -112,7 +107,6 @@ export const FavoriteRecipes = () => {
       </>
     )
   }
-
   return (
     <>
       <INTERFACE RENDERPAGE={<>
@@ -120,30 +114,14 @@ export const FavoriteRecipes = () => {
           {favoriteRecipes?.map((favorite, index) => {
             return (
               <Grid item xs={12} sm={6} md={4} lg={3}>
-                <Card
+                <Tag.FavoriteCard
                   key={index}
                   sx={{
                     width: isSmallScreen ? '100%' : 345,
-                    maxHeight: 'auto',
                     backgroundImage: `url(${!favorite?.recipeImage ? "https://www.biotecdermo.com.br/wp-content/uploads/2016/10/sem-imagem-10.jpg" : favorite?.recipeImage})`,
-                    backgroundSize: 'cover',
-                    backgroundRepeat: 'no-repeat',
-                    backgroundPosition: 'center',
                   }}
                 >
-                  <Box
-                    sx={{
-                      backgroundColor: 'rgba(225, 0, 0, 0.2)',
-                      backdropFilter: 'blur(3px)',
-                      transition: '.3s',
-                      color: 'white',
-                      '&:hover': {
-                        backgroundColor: 'transparent',
-                        backdropFilter: 'blur(0)',
-                        transition: '.3s',
-                      },
-                    }}
-                  >
+                  <Tag.FavoriteBox>
                     <CardHeader
                       avatar={
                         <Avatar
@@ -151,7 +129,7 @@ export const FavoriteRecipes = () => {
                           sx={{ bgcolor: "#212121", textTransform: 'uppercase' }}
                           aria-label="recipe"
                         >
-                          {favorite?.recipeTitle.charAt(0)}
+                          {favorite?.recipeTitle?.charAt(0)}
                         </Avatar>
                       }
                       action={
@@ -173,6 +151,8 @@ export const FavoriteRecipes = () => {
                       image={favorite?.recipeImage}
                       alt="Paella dish"
                     />
+
+
                     <CardContent sx={{ color: 'white' }}>
                       <Typography variant="body2">
                         {favorite?.recipeDescription}
@@ -188,14 +168,16 @@ export const FavoriteRecipes = () => {
                           <DeleteIcon />
                         </IconButton>
                       </Tooltip>
-                      <Tooltip title={`"Detalhar a receita de "${favorite?.recipeTitle}"`}>
-                        <IconButton
-                          size="small"
-                          sx={{ ml: 2, color: "#374957" }}
-                        >
-                          <DetailsIcon />
-                        </IconButton>
-                      </Tooltip>
+                      <Link href={`detailsRecipes/${favorite?.id}`}>
+                        <Tooltip title={`"Detalhar a receita de "${favorite?.id}"`}>
+                          <IconButton
+                            size="small"
+                            sx={{ ml: 2, color: "#374957" }}
+                          >
+                            <DetailsIcon />
+                          </IconButton>
+                        </Tooltip>
+                      </Link>
                       <ExpandMore
                         expand={expanded[index]}
                         onClick={() => handleExpandClick(index)}
@@ -211,8 +193,8 @@ export const FavoriteRecipes = () => {
                         <Typography paragraph>{favorite?.description}</Typography>
                       </CardContent>
                     </Collapse>
-                  </Box>
-                </Card>
+                  </Tag.FavoriteBox>
+                </Tag.FavoriteCard>
               </Grid>
             )
           })}
