+++ src/pages/INTERFACE/consts/links.jsx
@@ -61,7 +61,7 @@ export const links = [
   },
   {
     icon: <KitchenIcon />,
-    name: 'Recitas',
+    name: 'Receitas',
     onClick: 'recipes',
     children: [
       {
